+++ src/Adapters/LaravelMySQL/LaravelMySQLReuse.php
@@ -21,23 +21,30 @@ class LaravelMySQLReuse implements ReuseInterface
     /**
      * Insert details to the database to help identify if it can be reused or not.
      *
-     * @param string  $origDBName   The database that this test-database is for name.
-     * @param string  $snapshotHash The current snapshot-hash based on the database-building file content,
-     *                              pre-migration-imports, migrations and seeder-settings.
-     * @param boolean $reusable     Whether this database can be reused or not.
+     * @param string  $origDBName      The database that this test-database is for name.
+     * @param string  $sourceFilesHash The current source-files-hash based on the database-building file content.
+     * @param string  $scenarioHash    The current scenario-hash based on the pre-migration-imports, migrations and
+     *                                 seeder-settings.
+     * @param boolean $reusable        Whether this database can be reused or not.
      * @return void
      */
-    public function writeReuseData(string $origDBName, string $snapshotHash, bool $reusable)
-    {
+    public function writeReuseData(
+        string $origDBName,
+        string $sourceFilesHash,
+        string $scenarioHash,
+        bool $reusable
+    ) {
+
         $this->di->db->statement("DROP TABLE IF EXISTS `".Settings::REUSE_TABLE."`");
         $this->di->db->statement(
             "CREATE TABLE `".Settings::REUSE_TABLE."` ("
             ."`project_name` varchar(255), "
             ."`reuse_table_version` varchar(16), "
             ."`orig_db_name` varchar(255) NOT NULL, "
-            ."`snapshot_hash` varchar(255) NOT NULL, "
+            ."`source_files_hash` varchar(255) NOT NULL, "
+            ."`scenario_hash` varchar(255) NOT NULL, "
             ."`reusable` tinyint unsigned, "
-            ."`inside_transaction` tinyint unsigned "
+            ."`inside_transaction` tinyint unsigned"
             .")"
         );
         $this->di->db->insert(
@@ -45,16 +52,26 @@ class LaravelMySQLReuse implements ReuseInterface
                 ."`project_name`, "
                 ."`reuse_table_version`, "
                 ."`orig_db_name`, "
-                ."`snapshot_hash`, "
+                ."`source_files_hash`, "
+                ."`scenario_hash`, "
                 ."`reusable`, "
                 ."`inside_transaction`"
             .") "
-            ."VALUES (:projectName, :reuseTableVersion, :origDBName, :snapshotHash, :reusable, :insideTransaction)",
+            ."VALUES ("
+                .":projectName, "
+                .":reuseTableVersion, "
+                .":origDBName, "
+                .":sourceFilesHash, "
+                .":scenarioHash, "
+                .":reusable, "
+                .":insideTransaction"
+            .")",
             [
                 'projectName' => $this->config->projectName,
                 'reuseTableVersion' => Settings::REUSE_TABLE_VERSION,
                 'origDBName' => $origDBName,
-                'snapshotHash' => $snapshotHash,
+                'sourceFilesHash' => $sourceFilesHash,
+                'scenarioHash' => $scenarioHash,
                 'reusable' => (int) $reusable,
                 'insideTransaction' => false,
             ]
@@ -64,12 +81,13 @@ class LaravelMySQLReuse implements ReuseInterface
     /**
      * Check to see if the database can be reused.
      *
-     * @param string $snapshotHash The current snapshot-hash based on the database-building file content,
-     *                             pre-migration-imports, migrations and seeder-settings.
+     * @param string $sourceFilesHash The current source-files-hash based on the database-building file content.
+     * @param string $scenarioHash    The scenario-hash based on the pre-migration-imports, migrations and
+     *                                seeder-settings.
      * @return boolean
      * @throws AdaptBuildException When the database is owned by another project.
      */
-    public function dbIsCleanForReuse(string $snapshotHash): bool
+    public function dbIsCleanForReuse(string $sourceFilesHash, string $scenarioHash): bool
     {
         try {
             $rows = $this->di->db->select("SELECT * FROM `".Settings::REUSE_TABLE."` LIMIT 0, 1");
@@ -89,9 +107,19 @@ class LaravelMySQLReuse implements ReuseInterface
             );
         }
 
-        if (($reuseInfo->reuse_table_version != Settings::REUSE_TABLE_VERSION)
-        || ($reuseInfo->snapshot_hash != $snapshotHash)
-        || (!$reuseInfo->reusable)) {
+        if ($reuseInfo->reuse_table_version != Settings::REUSE_TABLE_VERSION) {
+            return false;
+        }
+
+        if ($reuseInfo->source_files_hash != $sourceFilesHash) {
+            return false;
+        }
+
+        if ($reuseInfo->scenario_hash != $scenarioHash) {
+            return false;
+        }
+
+        if (!$reuseInfo->reusable) {
             return false;
         }
 
@@ -112,15 +140,15 @@ class LaravelMySQLReuse implements ReuseInterface
      * Only removes databases that have reuse-info stored,
      * and that were for the same original database that this instance is for.
      *
-     * @param string|null $origDBName    The original database that this instance is for - will be ignored when null.
-     * @param string      $filesHash     The current files-hash based on the database-building file content.
-     * @param boolean     $detectOld     Remove old databases.
-     * @param boolean     $detectCurrent Remove new databases.
+     * @param string|null $origDBName      The original database that this instance is for - will be ignored when null.
+     * @param string      $sourceFilesHash The current files-hash based on the database-building file content.
+     * @param boolean     $detectOld       Detect old databases.
+     * @param boolean     $detectCurrent   Detect new databases.
      * @return string[]
      */
     public function findRelevantDatabases(
         $origDBName,
-        string $filesHash,
+        string $sourceFilesHash,
         bool $detectOld,
         bool $detectCurrent
     ): array {
@@ -137,7 +165,7 @@ class LaravelMySQLReuse implements ReuseInterface
             if ($this->isDatabaseRelevant(
                 $reuseInfo,
                 $origDBName,
-                $filesHash,
+                $sourceFilesHash,
                 $detectOld,
                 $detectCurrent
             )) {
@@ -150,17 +178,18 @@ class LaravelMySQLReuse implements ReuseInterface
     /**
      * Check to see if the given database is relevant.
      *
-     * @param stdClass|null $reuseInfo     The reuse info from the database.
-     * @param string|null   $origDBName    The original database that this instance is for - will be ignored when null.
-     * @param string        $filesHash     The current files-hash based on the database-building file content.
-     * @param boolean       $detectOld     Detect old databases.
-     * @param boolean       $detectCurrent Detect new databases.
+     * @param stdClass|null $reuseInfo       The reuse info from the database.
+     * @param string|null   $origDBName      The original database that this instance is for - will be ignored when
+     *                                       null.
+     * @param string        $sourceFilesHash The current files-hash based on the database-building file content.
+     * @param boolean       $detectOld       Detect old databases.
+     * @param boolean       $detectCurrent   Detect new databases.
      * @return boolean
      */
     private function isDatabaseRelevant(
         $reuseInfo,
         $origDBName,
-        string $filesHash,
+        string $sourceFilesHash,
         bool $detectOld,
         bool $detectCurrent
     ): bool {
@@ -177,13 +206,12 @@ class LaravelMySQLReuse implements ReuseInterface
             return false;
         }
 
+        // pick this up as "relevant" because it's obselete and should be replaced
         if ($reuseInfo->reuse_table_version != Settings::REUSE_TABLE_VERSION) {
             return true;
         }
 
-        $detectedFilesHash = $this->hasher->pickFileHashFromSnapshotHash($reuseInfo->snapshot_hash);
-        $filesHashMatched = ($detectedFilesHash == $filesHash);
-
+        $filesHashMatched = ($reuseInfo->source_files_hash == $sourceFilesHash);
         return ((($detectOld) && (!$filesHashMatched))
             || (($detectCurrent) && ($filesHashMatched)));
     }
@@ -198,8 +226,10 @@ class LaravelMySQLReuse implements ReuseInterface
     public function removeDatabase(string $database, bool $isOld = false): bool
     {
         $logTimer = $this->di->log->newTimer();
+
         $pdo = $this->di->db->newPDO();
         $success = $pdo->dropDatabase("DROP DATABASE IF EXISTS `".$database."`");
+
         $this->di->log->info('Removed '.($isOld ? 'old ' : '').'database: "'.$database.'"', $logTimer);
 
         return $success;