+++ src/Support/LaravelSupport.php
@@ -5,7 +5,7 @@ namespace CodeDistortion\Adapt\Support;
 use Orchestra\Testbench\TestCase;
 
 /**
- * Provides extra miscellaneous Laravel related support functionality
+ * Provides extra miscellaneous Laravel related support functionality.
  */
 class LaravelSupport
 {
@@ -16,8 +16,8 @@ class LaravelSupport
      */
     public static function isRunningInOrchestra(): bool
     {
-        $basePath = base_path();
-        $realpath = realpath('.');
+        $basePath = (string) base_path();
+        $realpath = (string) realpath('.');
         if (mb_strpos($basePath, $realpath) === 0) {
             $rest = mb_substr($basePath, mb_strlen($realpath));
             return (mb_substr($rest, 0, mb_strlen('/vendor/orchestra/')) == '/vendor/orchestra/');