+++ tests/Unit/DTO/ConfigDTOTest.php
@@ -7,7 +7,7 @@ use CodeDistortion\Adapt\DTO\ConfigDTO;
 use CodeDistortion\Adapt\Tests\PHPUnitTestCase;
 
 /**
- * Test the ConfigDTO class
+ * Test the ConfigDTO class.
  *
  * @phpcs:disable PSR1.Methods.CamelCapsMethodName.NotCamelCaps
  */
@@ -36,6 +36,10 @@ class ConfigDTOTest extends PHPUnitTestCase
             'database' => [
                 'method' => 'database',
                 'params' => ['database' => 'my_database'],
+            ],
+            'databaseModifier' => [
+                'method' => 'databaseModifier',
+                'params' => ['databaseModifier' => '_1'],
             ],
 
             'storageDir' => [